+++ src/ABHTTPServer.ts
@@ -1,19 +1,16 @@
-import { Socket } from "net";
-
-/*
-Simple HTTP Server Framework
-*/
+import {Socket} from 'net'
+import Express = require('express')
 
 /**
- * ABHTTPServer Abstract Class
- */
-export default abstract class ABHTTPServer {
+ * Simple HTTP Server Framework
+*/
+export abstract class ABHttpServer {
 
-  private DEBUG       = (process.env.AB_DEBUG === 'true') ? true : false
-  private express     = require('express')
-  private app         = this.express()
-  private httpServer  = null
-  private httpsServer = null
+  private DEBUG: boolean          = (process.env.AB_DEBUG === 'true') ? true : false
+  private app                     = Express()
+  private httpServer: any         = null
+  private httpsServer: any        = null
+  private httpHeaders: any        = {}  
 
   /**
    * Create the HTTP server
@@ -22,130 +19,140 @@ export default abstract class ABHTTPServer {
    */
   constructor(httpPort: number, httpsPort: number) {
 
-    this.DEBUG ? this.logDebug(`Constructor called with (${httpPort}, ${httpPort})`) : true
+    this.DEBUG ? this.logDebug(`Constructor called (${httpPort}, ${httpsPort})`) : true
 
     const http      = require('http')
     const https     = require('https')
     const fs        = require('fs')
     const path      = require('path')
   
-    /* Check constructor arguments */
+    // Check constructor arguments
     if (typeof arguments[0] !== 'number' || typeof arguments[1] !== 'number') {
-      this.DEBUG ? this.logDebug('Usage is ABHTTPServer(httpPort | 0, httpsPort | 0') : true
-      return (undefined)
+       throw new Error('ABHttpServer: Usage is ABHttpServer(httpPort | 0, httpsPort | 0')
     }
     if (arguments[0] < 0 || arguments[0] > 65535 || arguments[1] < 0 || arguments[1] > 65535) {
-      this.DEBUG ? this.logDebug('Both port arguments must be between 0 and 65535') : true
-      return (undefined)
+      throw new Error('ABHttpServer: Both port arguments must be between 0 and 65535')
     }
     if (arguments[0] % 1 !== 0 || arguments[1] % 1 !== 0) {
-      this.DEBUG ? this.logDebug('Both port arguments must have integer values') : true
-      return (undefined)
+      throw new Error('ABHttpServer: Both port arguments must have integer values')
     }
     if (arguments[0] === arguments[1]) {
-      this.DEBUG ? this.logDebug('Both ports must not be equal') : true
-      return (undefined)
+      throw new Error('ABHttpServer: Both ports must not be equal')
     }
     if (arguments[0] === 0 && arguments[1] === 0) {
-      this.DEBUG ? this.logDebug('At least one port must be non-zero') : true
-      return (undefined)
+      throw new Error('ABHttpServer: At least one port must be non-zero')
     }
     
-    /* Start the HTTP server */
+    // Start the HTTP server
     if (httpPort != 0) {
+
       this.DEBUG ? this.logDebug(`Creating HTTP server on port ${httpPort}`) : true
+
       this.httpServer = http.createServer(this.app)
       this.startServer(this.httpServer, httpPort, false)
     }
 
-    /* Start the HTTPS server */
+    // Start the HTTPS server
     if (httpsPort != 0) {
+
       this.DEBUG ? this.logDebug(`Creating HTTPS server on port ${httpsPort}`) : true
+
       var httpsOptions = {
         key:  fs.readFileSync(path.join(__dirname, 'x509-servercert.pem')),
-        cert: fs.readFileSync(path.join(__dirname, 'x509-certchain.pem'))
+        cert: fs.readFileSync(path.join(__dirname, 'x509-certchain.pem')),
         /* Add "allowHTTP1: true" when using the http2 module */
       }
+
       this.httpsServer = https.createServer(httpsOptions, this.app)
       this.startServer(this.httpsServer, httpsPort, true)
     }
 
-    /* Handle all HTTP requests */
-    this.app.all('*', (request, response) => {
+    // Do not send Express HTTP header 'x-powered-by'
+    this.app.disable('x-powered-by')
+
+    // Handle all HTTP requests
+    this.app.all('*', (request: Express.Request, response: Express.Response) => {
+      
       this.DEBUG ? this.logDebug(`${request.protocol.toUpperCase()} ${request.httpVersion} ${request.method} ${request.url}`) : true
 
       switch (request.method) {
-        /* Method GET */
         case 'GET': {
           this.get(request.url, request, response)
           break
         }
           
-        /* Method HEAD */
         case 'HEAD': {
           this.head(request.url, request, response)
           break
         }
           
-        /* Method PUT */
         case 'PUT': {
-          let postData = ''
-          request.on('data', (dataChunk) => {
-            postData += dataChunk
-          })
-          request.on('end', () =>  {
-            this.DEBUG ? this.logDebug(`Received PUT data size: ${postData.length} bytes`) : true
-          })
-          this.put(request.url, postData, request, response)
-          break
-        }
-        
-        /* Method POST */
-        case 'POST': {
-          let postData = ''
-          request.on('data', (dataChunk) => {
-            postData += dataChunk
+          let requestData = ''
+          request.on('data', (dataChunk: string) => {
+            requestData += dataChunk
           })
           request.on('end', () => {
-            this.DEBUG ? this.logDebug(`Received POST data size: ${postData.length} bytes`) : true
+            this.DEBUG ? this.logDebug(`PUT data received: ${requestData.length} bytes`) : true
+            this.put(request.url, requestData, request, response)
           })
-          this.post(request.url, postData, request, response)
           break
         }
         
-        /* Method DELETE */
+        case 'POST': {
+          let requestData = ''
+          request.on('data', (dataChunk: string) => {
+            requestData += dataChunk
+          })
+          request.on('end', () => {
+            this.DEBUG ? this.logDebug(`POST data received: ${requestData.length} bytes`) : true
+            this.post(request.url, requestData, request, response)
+          })
+          break
+        }
+        
         case 'DELETE': {
-          this.delete(request.url, request, response)
+          let requestData = ''
+          request.on('data', (dataChunk: string) => {
+            requestData += dataChunk
+          })
+          request.on('end', () => {
+            this.DEBUG ? this.logDebug(`DELETE data received: ${requestData.length} bytes`) : true
+            this.delete(request.url, requestData, request, response)
+          })
           break
         }
           
-        /* Method CONNECT */
         case 'CONNECT': {
           this.connect(request.url, request, response)
           break
         }
           
-        /* Method TRACE */
         case 'TRACE': {
           this.trace(request.url, request, response)
           break
         }
           
-        /* Method PATCH */
         case 'PATCH': {
-          this.patch(request.url, request, response)
+          let requestData = ''
+          request.on('data', (dataChunk: string) => {
+            requestData += dataChunk
+          })
+          request.on('end', () => {
+            this.DEBUG ? this.logDebug(`PATCH data received: ${requestData.length} bytes`) : true
+            this.patch(request.url, requestData, request, response)
+          })
+
           break
         }
           
-        /* Method OPTIONS */
         case 'OPTIONS': {
           this.options(request.url, request, response)
           break
         }
         
         default: {
-          this.DEBUG ? this.logDebug(`Unhandled HTTP method ${request.method}`) : true 
-          this.sendText(response, `Unhandled HTTP method <${request.method}>`)
+          this.DEBUG ? this.logDebug(`Unsupported HTTP method ${request.method}`) : true 
+          this.sendText(response, `ABHttpServer: Unsupported HTTP method <${request.method}>`, 501)
         }
       }
     });
@@ -157,9 +164,9 @@ export default abstract class ABHTTPServer {
    * @param port    TCP/IP port number
    * @param secure  SSL/TLS flag
    */
-  private startServer(server, port: number, secure: boolean) : void {
+  private startServer(server: any, port: number, secure: boolean): void {
    
-    /* Establish net.Server event handlers */
+    // Establish net.Server event handlers
     server.on('error', (error: Error) => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ` (net.Server) server received error: ${error}`) : true
     })
@@ -167,12 +174,12 @@ export default abstract class ABHTTPServer {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ` (net.Server) server is in listen mode`) : true
     })
 
-    /* Establish http.Server event handlers */
-    server.on('checkContinue', (request, response) => {
+    // Establish http.Server event handlers
+    server.on('checkContinue', (request: Express.Request, response: Express.Response) => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ' (http.Server) server received <HTTP 100 continue>') : true
       response.writeContinue()
     })
-    server.on('checkExpectation', (request, response) => {
+    server.on('checkExpectation', (request: Express.Request, response: Express.Response) => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ' (http.Server) server received HTTP expect header') : true
     })
     server.on('clientError', (err: Error, socket: Socket) => {
@@ -182,40 +189,40 @@ export default abstract class ABHTTPServer {
     server.on('close', () => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ' (http.Server) server closed') : true
     })
-    server.on('connect', (request, socket: Socket, head: Buffer) => {
+    server.on('connect', (request: Express.Request, socket: Socket, head: Buffer) => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ' (http.Server) server accepted connection') : true
     })
     server.on('connection', (socket: Socket) => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ' (http.Server) server established connection') : true
     })
-    server.on('request', (request, response) => {
+    server.on('request', (request: Express.Request, response: Express.Response) => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ' (http.Server) server received client request') : true
     })
-    server.on('upgrade', (request, socket: Socket, head: Buffer) => {
+    server.on('upgrade', (request: Express.Request, socket: Socket, head: Buffer) => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ' (http.Server) server received upgrade request') : true
     })
 
-    /* Establish tls.server event handlers */
-    server.on('newSession', (sessionId, sessionData, callback) => {
+    // Establish tls.server event handlers
+    server.on('newSession', (sessionId: any, sessionData: any, callback: Function) => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ' (tls.Server) server started TLS session') : true
       callback()
     })
-    server.on('OCSPRequest', (certificate, issuer, callback) => {
+    server.on('OCSPRequest', (certificate: any, issuer: any, callback: Function) => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ' (tls.Server) server received certificate status request') : true
       callback()
     })
-    server.on('resumeSession', (sessionId, callback) => {
+    server.on('resumeSession', (sessionId: any, callback: Function) => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ' (tls.Server) server received TLS resume session status request') : true
       callback()
     })
-    server.on('secureConnection', (tlsSocket) => {
+    server.on('secureConnection', (tlsSocket: any) => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ' (tls.Server) completed TLS handshaking process') : true
     })
-    server.on('tlsClientError', (exception, tlsSocket) => {
+    server.on('tlsClientError', (exception: Error, tlsSocket: any) => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ' (tls.Server) received an error before successful connection') : true
     })
 
-    /* Start the server */
+    // Start the server
     server.listen(port, () => {
       this.DEBUG ? this.logDebug('HTTP' + (secure ? 'S' : '') + ` server started on port ${port}`) : true
     })
@@ -225,19 +232,19 @@ export default abstract class ABHTTPServer {
   * Write debugging data to the console
   * @param message Debug message to be written
   */
-  private logDebug(message) {
+  private logDebug(message: string) {
     if (this.DEBUG) {
       
       var date = new Date()
       var timestamp = date.getFullYear() + '-' +
-        this.dateTimePad((date.getMonth() + 1), 2) + '-' +
-        this.dateTimePad(date.getDate(), 2) + ' ' +
-        this.dateTimePad(date.getHours(), 2) + ':' +
-        this.dateTimePad(date.getMinutes(), 2) + ':' +
-        this.dateTimePad(date.getSeconds(), 2) + '.' +
-        this.dateTimePad(date.getMilliseconds(), 3)
+        this.dateTimePad((date.getMonth() + 1).toString(), 2) + '-' +
+        this.dateTimePad(date.getDate().toString(), 2) + ' ' +
+        this.dateTimePad(date.getHours().toString(), 2) + ':' +
+        this.dateTimePad(date.getMinutes().toString(), 2) + ':' +
+        this.dateTimePad(date.getSeconds().toString(), 2) + '.' +
+        this.dateTimePad(date.getMilliseconds().toString(), 3)
       
-      console.log(`${timestamp} ABHTTPSERVER Debug: ${message}`)
+      console.debug(`${timestamp} ABHttpServer: ${message}`)
     }
   }
 
@@ -246,7 +253,7 @@ export default abstract class ABHTTPServer {
    * @param value 
    * @param digits 
    */
-  private dateTimePad(value, digits) : number {
+  private dateTimePad(value: string, digits: number) : string {
     let number = value
     while (number.toString().length < digits) {
       number = `0${number}`
@@ -259,14 +266,14 @@ export default abstract class ABHTTPServer {
    */
   terminate(): void {
 
-    /* Stop HTTP server */
-    if (this.httpServer !== null) {
+    this.DEBUG ? this.logDebug('Invoked method terminate()') : true
+
+    if (this.httpServer) {
       this.httpServer.close()
       this.httpServer = null
     }
 
-    /* Stop HTTPS server */
-    if (this.httpsServer !== null) {
+    if (this.httpsServer) {
       this.httpsServer.close()
       this.httpsServer = null
     }
@@ -282,34 +289,110 @@ export default abstract class ABHTTPServer {
 
   /**
    * Sends HTML data to the client
-   * @param response  
-   * @param text      HTML to be sent
+   * @param response    Express.Response object  
+   * @param text        HTML to be sent
+   * @param httpStatus  HTTP Status code (defaults to 200)
    */
-  sendHTML(response: any, text: string) {
-    response.setHeader('Content-Type', 'text/html')
-    response.end(text)
+  sendHTML(response: Express.Response, text: string, httpStatus: number = 200) {
+    this.sendData(response, 'text/html', text, httpStatus)
   }
 
   /**
    * Sends plain text to the client
    * @param response  
    * @param text      Text to be sent
+   * @param httpStatus  HTTP Status code (defaults to 200)
    */
-  sendText(response: any, text: string) {
-    response.setHeader('Content-Type', 'text/plain')
+  sendText(response: Express.Response, text: string, httpStatus: number = 200) {
+    this.sendData(response, 'text/plain', text, httpStatus)
+  }
+
+  /**
+   * Set HTTP headers to be added to every response
+   * 
+   * Example:
+   * this.setHeaders({'Access-Control-Allow-Origin': '*',
+   *                  'Access-Control-Allow-Methods': 'GET, POST, DELETE, PUT' })
+   * 
+   * @param httpHeaders HTTP Headers to be added
+   */
+  setHeaders(httpHeaders: any) {
+
+    this.DEBUG ? this.logDebug(`Invoked method setHeaders(${httpHeaders})`) : true
+    this.httpHeaders = httpHeaders
+  }
+
+  /**
+   * Writes the data with HTTP header
+   * @param response      Express.Response
+   * @param mimeType      HTTP Content-Type
+   * @param text          Data to be written
+   * @param httpStatus    HTTP Status code
+   */
+  private sendData(response: Express.Response, mimeType: string, text: string, httpStatus: number) {
+
+    this.DEBUG ? this.logDebug(`Sending HTTP status ${httpStatus} with ${text.length} bytes of ${mimeType} data`) : true
+
+    // Send additional HTTP headers
+    if (!this.httpHeaders) {
+      for (let key in this.httpHeaders) {
+        response.set(key, this.httpHeaders[key])
+      }
+    }
+
+    response.writeHead(httpStatus, { 'Content-Type': mimeType })
     response.end(text)
   }
 
-  /* Method prototypes to be overwritten/implemented in subclass */
-  clientConnect(socket) { }
-  clientError(err, socket) { }
-  get(url: string, request: any, response: any) { }
-  put(url: string, data: string, request: any, response: any) { }
-  post(url: string, data: string, request: any, response: any) { }
-  options(url: string, request: any, response: any) { }
-  head(url: string, request: any, response: any) { }
-  delete(url: string, request: any, response: any) { }
-  connect(url: string, request: any, response: any) { }
-  trace(url: string, request: any, response: any) { }
-  patch(url: string, request: any, response: any) { }
+  /**
+   * Send HTTP 'Not Implemented' Error
+   * @param method 
+   * @param response Express.response
+   */
+  private notImplementedError(method: string, response: Express.Response) {
+
+    this.DEBUG ? this.logDebug(`HTTP method ${method} not implemented by subclass`) : true
+    response.status(501).end(`HTTP method ${method} is not supported`)
+  }
+
+  // Method prototypes to be overwritten in subclass
+  clientConnect(socket: Socket) { }
+
+  clientError(err: Error, socket: Socket) { }
+
+  get(url: string, request: Express.Request, response: Express.Response) { 
+    this.notImplementedError('GET', response)
+  }
+  
+  put(url: string, data: string, request: Express.Request, response: Express.Response) {
+    this.notImplementedError('PUT', response)
+  }
+
+  post(url: string, data: string, request: Express.Request, response: Express.Response) {
+    this.notImplementedError('POST', response)
+  }
+
+  options(url: string, request: Express.Request, response: Express.Response) { 
+    this.notImplementedError('OPTIONS', response)
+  }
+
+  head(url: string, request: Express.Request, response: Express.Response) { 
+    this.notImplementedError('HEAD', response)
+  }
+
+  delete(url: string, data: string, request: Express.Request, response: Express.Response) {
+    this.notImplementedError('DELETE', response)
+  }
+
+  connect(url: string, request: Express.Request, response: Express.Response) {
+    this.notImplementedError('CONNECT', response)
+  }
+
+  trace(url: string, request: Express.Request, response: Express.Response) {
+    this.notImplementedError('TRACE', response)
+  }
+
+  patch(url: string, data: string, request: Express.Request, response: Express.Response) {
+    this.notImplementedError('PATCH', response)
+  }
 }