+++ src/test.js
@@ -1,9 +1,12 @@
 "use strict";
-/* ABHTTPServer Test */
+/* ABHttpServer Test */
 var __extends = (this && this.__extends) || (function () {
-    var extendStatics = Object.setPrototypeOf ||
-        ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-        function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+    var extendStatics = function (d, b) {
+        extendStatics = Object.setPrototypeOf ||
+            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
+            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
+        return extendStatics(d, b);
+    }
     return function (d, b) {
         extendStatics(d, b);
         function __() { this.constructor = d; }
@@ -11,7 +14,7 @@ var __extends = (this && this.__extends) || (function () {
     };
 })();
 exports.__esModule = true;
-var ABHTTPServer_1 = require("./ABHTTPServer");
+var ABHttpServer_1 = require("./ABHttpServer");
 var MyServer = /** @class */ (function (_super) {
     __extends(MyServer, _super);
     function MyServer() {
@@ -21,5 +24,5 @@ var MyServer = /** @class */ (function (_super) {
         this.sendText(response, "Hello - The URL sent was " + url);
     };
     return MyServer;
-}(ABHTTPServer_1["default"]));
+}(ABHttpServer_1.ABHttpServer));
 var myServer = new MyServer(8080, 8081);