+++ HISTORY.md
@@ -1,5 +1,14 @@
 # Revision History
 
+## Version 1.0.0 / 05-Aug-2018
+
+* Change: Throw Error() in constructor for invalid parameters
+* Change: Disable sending the Express HTTP header 'x-powered-by'
+* Change: Overall code cleanup
+* Add: Return HTTP status 501 for HTTP methods not implemented by a user method
+* Add: Include generated definition files ABHttpServer.d.ts and test.d.ts
+* Add: Method setHeaders() to add HTTP headers to every response
+
 ## Version 0.2.0 / 20-Jul-2018
 
 * Change: Corrected error in sample code test.ts